interfaces in Go
interfaces in Go are a bit differently implemented
but at the end of the day it's really just a way of saying that a struct should
have a few methods.

Then in a different function, you can use a variable with as type the interface
to abstract away the different, but same-named, methods separate structs might have.